Modify Header and Gallery components to handle changes in image state
Currently set up to assign a default image to a currentImage in state. currentImage will be modified depending on what image you are toggling through (will need to properly set this up once API is connected). currentImage state is kept in the parent Header.jsx component so that when the gallery is reloaded, the currentImage is the one last viewed.